Refactor trace writer facade module
